streamlit_app/requirements.txt
numpy==2.2.3
pandas==1.5.3
Pillow==11.1.0
Requests==2.32.3
sentence_transformers==3.4.1
streamlit==1.39.0
streamlit_extras==0.5.5
